+++ woodpecker/stores/index.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from woodpecker.fixes.identifiers import IdentifierResolver
 
-from ..plans.models import FixPlan
+from ..fix_plans.models import FixPlan
 
 
 class FixPlanIndex: